Set and check reboot scripts
These scripts are meant to be called during a series of 'sub' policies
(policies called during the execution of other policies). The set
script marks that the system will need a reboot, and the check script,
running at the end of the 'master' policy, displays a prompt to the
user and reboots after a set amount of time.