A wireless redeployment script
I've been deploying a machine wireless authentication (using machine AD
credentials) to my deployment. I've had a number of situations where I
needed to redeploy the wireless profile, and my old method of doing
this was completely removing and reenrolling the mdm profile of the
Mac, but this script will at the Mac that runs it to an 'exclusion'
group which will remove the profile, and once it detects the profile is
gone it will remove the mac from the exclusion group. Unfortunately
only one Mac would be able to run this at a time, but this method means
I won't have to install a new openssl on my Macs to get python working
with SSL, so I thought it was worth it!